+++ src/messages/components/WallConversation.vue
@@ -1,5 +1,5 @@
 <template>
-  <div v-if="data">
+  <div>
     <QInfiniteScroll
       :handler="maybeFetchPast"
     >
@@ -8,10 +8,10 @@
       >
         <template v-if="hasLoaded">
           <NotificationToggle
+            class="actionButton hoverScale"
             :muted="data.muted"
             :is-participant="data.isParticipant"
             :user="user"
-            class="actionButton hoverScale"
             @set="setNotifications"
           />
           <ConversationCompose
@@ -53,11 +53,7 @@
             @openThread="$emit('openThread', message)"
           />
         </template>
-        <div
-          v-if="data.fetchStatus.pending || data.fetchPastStatus.pending"
-          style="width: 100%; text-align: center">
-          <QSpinnerDots :size="40"/>
-        </div>
+        <KSpinner v-show="!data || data.fetchStatus.pending || data.fetchPastStatus.pending" />
       </QList>
     </QInfiniteScroll>
   </div>
@@ -67,10 +63,10 @@
 import ConversationMessage from './ConversationMessage'
 import ConversationCompose from './ConversationCompose'
 import NotificationToggle from './NotificationToggle'
+import KSpinner from '@/utils/components/KSpinner'
 import {
   QBtn,
   QInfiniteScroll,
-  QSpinnerDots,
   QList,
   QAlert,
 } from 'quasar'
@@ -81,9 +77,9 @@ export default {
     ConversationMessage,
     ConversationCompose,
     NotificationToggle,
+    KSpinner,
     QBtn,
     QInfiniteScroll,
-    QSpinnerDots,
     QList,
     QAlert,
   },