+++ src/pickups/datastore/index.js
@@ -1,4 +1,5 @@
 import pickups from '@/pickups/datastore/pickups'
+import pickupPlugin from '@/pickups/datastore/pickupPlugin'
 import pickupSeries from '@/pickups/datastore/pickupSeries'
 
 export default {
@@ -7,5 +8,6 @@ export default {
     pickupSeries,
   },
   plugins: [
+    pickupPlugin,
   ],
 }