+++ src/users/datastore/users.js
@@ -8,6 +8,7 @@ import router from '@/base/router'
 function initialState () {
   return {
     entries: {},
+    activeUserProfileId: null,
     activeUserProfile: null,
     resetPasswordSuccess: false,
     resendVerificationCodeSuccess: false,
@@ -59,7 +60,7 @@ export default {
       }
     },
     activeUserId: state => state.activeUserProfile && state.activeUserProfile.id,
-    ...metaStatuses(['signup', 'requestResetPassword', 'resetPassword', 'resendVerificationCode', 'requestDeleteAccount']),
+    ...metaStatuses(['signup', 'requestResetPassword', 'resetPassword', 'resendVerificationCode', 'requestDeleteAccount', 'fetch']),
     resetPasswordSuccess: state => state.resetPasswordSuccess,
     resendVerificationCodeSuccess: state => state.resendVerificationCodeSuccess,
   },
@@ -99,20 +100,27 @@ export default {
       },
     }),
 
-    async selectUser ({ commit }, { userId }) {
-      try {
-        commit('setProfile', await users.getProfile(userId))
-      }
-      catch (error) {
+    ...withMeta({
+      async selectUser ({ commit }, { userId }) {
         try {
-          commit('setProfile', await users.getInfo(userId))
+          commit('setProfile', await users.getProfile(userId))
         }
         catch (error) {
-          const data = { translation: 'PROFILE.INACCESSIBLE_OR_DELETED' }
-          throw createRouteError(data)
+          try {
+            commit('setProfile', await users.getInfo(userId))
+          }
+          catch (error) {
+            const data = { translation: 'PROFILE.INACCESSIBLE_OR_DELETED' }
+            throw createRouteError(data)
+          }
         }
-      }
-    },
+      },
+    }, {
+      findId: ({ userId }) => userId,
+      setCurrentId: ({ commit }, { userId }) => commit('setProfileId', userId),
+      getCurrentId: ({ state }) => state.activeUserProfileId,
+    }),
+
     clearSelectedUser ({ commit }) {
       commit('setProfile', null)
     },
@@ -149,6 +157,9 @@ export default {
   mutations: {
     setProfile (state, userProfile) {
       state.activeUserProfile = userProfile
+    },
+    setProfileId (state, id) {
+      state.activeUserProfileId = id
     },
     update (state, users) {
       for (const user of users) {