+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Please document your changes in this format:
 - Allow non-collectors to read and write in pickup chat @tiltec
 - Pickups from non-favorited stores won't show up in the "available pickup" in the group wall and don't send emails about upcoming pickups @tiltec
 - Can now unsubscribe from conversations (except private conversation and threads) @tiltec
+- Show store name instead of "Current store" in sidenav @tiltec
 
 ## [6.5.0] - 2019-02-03
 ### Added